Updated error handling so when a Broker fails and a new leader is identified the Storm/Trident side doesn't crash.
However this is way less than ideal. Because it isn't clear what will happen if we change the data in the Partition object from Storm we are playing games under the covers to map the current Broker to data in the Partition object.

Need help from Nathan to understand the impact.

+++ storm-kafka/src/jvm/storm/kafka/BrokerData.java
@@ -4,10 +4,10 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class BrokerData {
-    private long port;
+    private int port;
     private List<Integer> partitions;
 
-    public BrokerData(long a_port) {
+    public BrokerData(int a_port) {
         port = a_port;
         partitions = new ArrayList<Integer>();
     }
@@ -15,7 +15,7 @@ public class BrokerData {
         partitions.add(a_partition);
     }
 
-    public long getPort() {
+    public int getPort() {
         return port;
     }
     public List<Integer> getPartitions() {
